mark as read option added for notification

+++ Backend/controllers/notification.js
@@ -9,14 +9,18 @@ const getNotifications = asyncWrapper(async (req, res, next) => {
     const { username } = req.body;
     const [result] = await db.execute(`SELECT * FROM notifications WHERE celebrity='${username}'`)
     // console.log("noties", result)
-    res.status(StatusCodes.OK).json({result:result})
+    res.status(StatusCodes.OK).json({ result: result })
 
 })
 
 
 const deleteNotification = asyncWrapper(async (req, res, next) => {
     const _id = req.params.id;
-    await db.execute(`DELETE FROM notifications WHERE id ="${_id}"`)
+    let result = null;
+    if (_id !== "-1") {
+        result = await db.execute(`DELETE FROM notifications WHERE id ="${_id}"`)
+    }
+    res.status(StatusCodes.OK).json({ result: result })
 })
 
-module.exports=  {getNotifications,deleteNotification}
+module.exports = { getNotifications, deleteNotification }
